database refactor

+++ src/main/java/com/ustudent/resquod/repository/PositionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface PositionRepository extends JpaRepository<Position,Integer> {
+public interface PositionRepository extends JpaRepository<Position,Long> {
 
     Optional<Position> findByNumberOfPosition(Integer numberOfPosition);
 }
