+++ src/main/java/com/ustudent/resquod/repository/UserRepository.java
@@ -11,8 +11,8 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-    Optional<User> findById(Integer id);
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findById(Long id);
 
     Optional<User> findByEmail(String email);
